+++ lib/auth/data/repository_impl/auth_repository_impl.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:devlink_mobile_app/auth/data/data_source/auth_data_source.dart';
 import 'package:devlink_mobile_app/auth/data/dto/timer_activity_dto.dart';
-import 'package:devlink_mobile_app/auth/data/mapper/user_mapper.dart';
+import 'package:devlink_mobile_app/auth/data/mapper/member_mapper.dart';
 import 'package:devlink_mobile_app/auth/domain/model/member.dart';
 import 'package:devlink_mobile_app/auth/domain/model/terms_agreement.dart';
 import 'package:devlink_mobile_app/auth/domain/repository/auth_repository.dart';
@@ -10,9 +10,9 @@ import 'package:devlink_mobile_app/core/auth/auth_state.dart';
 import 'package:devlink_mobile_app/core/config/app_config.dart';
 import 'package:devlink_mobile_app/core/result/result.dart';
 import 'package:devlink_mobile_app/core/utils/api_call_logger.dart';
+import 'package:devlink_mobile_app/core/utils/app_logger.dart';
 import 'package:devlink_mobile_app/core/utils/exception_mappers/auth_exception_mapper.dart';
 import 'package:devlink_mobile_app/core/utils/messages/auth_error_messages.dart';
-import 'package:devlink_mobile_app/core/utils/app_logger.dart';
 import 'package:devlink_mobile_app/core/utils/privacy_mask_util.dart';
 import 'package:devlink_mobile_app/notification/service/fcm_token_service.dart';
 import 'package:firebase_auth/firebase_auth.dart';
